+++ Common/src/main/java/fuzs/puzzleslib/api/util/v1/ShapesHelper.java
@@ -2,6 +2,7 @@ package fuzs.puzzleslib.api.util.v1;
 
 import com.google.common.collect.Maps;
 import net.minecraft.core.Direction;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.joml.Quaternionf;
@@ -9,6 +10,8 @@ import org.joml.Vector3d;
 
 import java.util.EnumMap;
 import java.util.Map;
+import java.util.function.IntFunction;
+import java.util.stream.IntStream;
 
 /**
  * A simple helper class for {@link VoxelShape}, mainly used for applying rotations.
@@ -123,5 +126,66 @@ public final class ShapesHelper {
                 Math.max(startX, endX),
                 Math.max(startY, endY),
                 Math.max(startZ, endZ));
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape[] boxes(int endInclusive, IntFunction<VoxelShape> voxelShapeFactory) {
+        return IntStream.rangeClosed(0, endInclusive).mapToObj(voxelShapeFactory).toArray(VoxelShape[]::new);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape cube(double size) {
+        return cube(size, size, size);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape cube(double sizeX, double sizeY, double sizeZ) {
+        double halfY = sizeY / 2.0;
+        return column(sizeX, sizeZ, 8.0 - halfY, 8.0 + halfY);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape column(double sizeXZ, double minY, double maxY) {
+        return column(sizeXZ, sizeXZ, minY, maxY);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape column(double sizeX, double sizeZ, double minY, double maxY) {
+        double halfX = sizeX / 2.0;
+        double halfZ = sizeZ / 2.0;
+        return Block.box(8.0 - halfX, minY, 8.0 - halfZ, 8.0 + halfX, maxY, 8.0 + halfZ);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape boxZ(double sizeXY, double minZ, double maxZ) {
+        return boxZ(sizeXY, sizeXY, minZ, maxZ);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape boxZ(double sizeX, double sizeY, double minZ, double maxZ) {
+        double halfY = sizeY / 2.0;
+        return boxZ(sizeX, 8.0 - halfY, 8.0 + halfY, minZ, maxZ);
+    }
+
+    /**
+     * Copied from Minecraft 26.1.
+     */
+    public static VoxelShape boxZ(double sizeX, double minY, double maxY, double minZ, double maxZ) {
+        double halfX = sizeX / 2.0;
+        return Block.box(8.0 - halfX, minY, minZ, 8.0 + halfX, maxY, maxZ);
     }
 }