+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.1.43-1.21.1] - 2025-05-21
+
+### Added
+
+- Add `TransformingForwardingList`
+- Add some helper methods to `ShapesHelper`
+- Add some helper methods to `AbstractRecipeProvider`
+
 ## [v21.1.42-1.21.1] - 2025-05-20
 
 ### Added